+++ components/mysql/actions/common/utils.mjs
@@ -1,0 +1,34 @@
+import mysql from "../../mysql.app.mjs";
+
+async function getColumnProps(table) {
+  const props = {};
+  const columns = await this.mysql.listColumnNames(table);
+  for (const column of columns) {
+    props[column] = {
+      type: "string",
+      label: column,
+      optional: true,
+    };
+  }
+  return props;
+}
+
+async function getColumnAndValueArrays(table) {
+  const columns = [];
+  const values = [];
+  const columnNames = await mysql.methods.listColumnNames(table);
+  for (const column of columnNames) {
+    if (this[column]) {
+      columns.push(column);
+      values.push(this[column]);
+    }
+  }
+  return {
+    columns,
+    values,
+  };
+}
+export default {
+  getColumnProps,
+  getColumnAndValueArrays,
+};
